+++ octree_web_viewer/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod backend;
 pub mod backend_error;
+pub mod state;
 pub mod utils;
